requirements.txt
annotated-types==0.7.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
deepl==1.18.0
google-ai-generativelanguage==0.6.9
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.1
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
httplib2==0.22.0
idna==3.9
proto-plus==1.24.0
protobuf==5.28.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.1
pydantic_core==2.23.3
pyparsing==3.1.4
pyTelegramBotAPI==4.22.1
requests==2.32.3
rsa==4.9
telebot==0.0.5
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3

openai~=0.28.0
python-dotenv~=1.0.1
langchain~=0.3.0